joey/ server follows route done & tested

+++ Server/routes/follows.js
@@ -4,27 +4,38 @@ GROUP 1: Amine Bensalem, Douglas MacKrell, Savita Madray, Joseph P. Pasaoa
 */
 
 
-/* MODULE INITS */
+/* IMPORTS */
 const express = require('express');
-const router = express.Router();
-// Queries
-const users = require('../queries/authentication');
+    const router = express.Router();
+    
+// local
+const { authenticateUser } = require('../queries/authentication.js'); // for authentication
+const { getUserById } = require('../queries/users.js'); // for checking if user exists after no results
 const { 
   getFollows,
   getFollowers,
+  checkFollowExists,
   createFollow,
   deleteFollow
-} = require('../queries/follows');
+} = require('../queries/follows.js');
 
 
 /* HELPERS */
-const handleError = (res, error) => {
-  console.log(error);
-  res.status(500);
+const handleError = (res, error, code) => {
+  console.log(code ? `error(fe): ${error}` : `error(be): ${error}`);
+  res.status(code || 500);
   res.json({
       status: "fail",
-      message: "error: backend issue",
+      message: code ? `error: ${error}` : `error(backend): ${error}`,
       payload: null
+  });
+}
+
+const handleSuccess = (res, path, data) => {
+  res.json({
+    status: "success",
+    message: `${path} data retrieved`,
+    payload: data.length === 1 ? data[0] : data
   });
 }
 
@@ -38,12 +49,9 @@ const checkIdParams = (req) => {
   }
   // COMPILE MESSAGE
   if (problems.length) {
-    if (problems.length >= 2) {
-      problems[problems.length - 1] = "and " + problems[problems.length - 1];
-      problems = problems.join(' ');
-    } else {
-      problems = problems.join('');
-    }
+    problems.length == 2
+      ? problems = problems.join(' and ')
+      : problems = problems.join('');
     return problems;
   }
   return false;
@@ -54,15 +62,21 @@ const checkIdParams = (req) => {
 // getFollows: get all others the user is following
 router.get("/:currUserId", async (req, res) => {
     if (!req.params.currUserId || isNaN(parseInt(req.params.currUserId.trim()))) {
-      handleError(res, 'invalid currUserId parameter');
+      handleError(res, 'invalid currUserId parameter', 400);
     } else {
+      const currUserId = parseInt(req.params.currUserId.trim());
       try {
-        const follows = await getFollows(parseInt(req.params.currUserId.trim()));
-        res.json({
-            status: "success",
-            message: "followings retrieved",
-            payload: follows
-        });
+        const follows = await getFollows(currUserId);
+        if (follows.length === 0) {
+          const userExists = await getUserById(currUserId);
+          if (userExists === 'no match') {
+            handleError(res, 'user does not exist', 400);
+          } else {
+            handleSuccess(res, 'follows', follows);
+          }
+        } else {
+          handleSuccess(res, 'follows', follows);
+        }
       } catch (err) {
         handleError(res, err);
       }
@@ -70,17 +84,23 @@ router.get("/:currUserId", async (req, res) => {
 });
 
 // getFollowers: get all following current user
-router.get("/whofollows/:currUserId", async (req, res) => {
+router.get("/followers/:currUserId", async (req, res) => {
     if (!req.params.currUserId || isNaN(parseInt(req.params.currUserId.trim()))) {
-      handleError(res, 'invalid currUserId parameter');
+      handleError(res, 'invalid currUserId parameter', 400);
     } else {
+      const currUserId = parseInt(req.params.currUserId.trim());
       try {
-        const followers = await getFollowers(parseInt(req.params.currUserId.trim()));
-        res.json({
-            status: "success",
-            message: "followers retreived",
-            payload: followers
-        });
+        const followers = await getFollowers(currUserId);
+        if (followers.length === 0) {
+          const userExists = await getUserById(currUserId);
+          if (userExists === 'no match') {
+            handleError(res, 'user does not exist', 400);
+          } else {
+            handleSuccess(res, 'followers', followers);
+          }
+        } else {
+          handleSuccess(res, 'followers', followers);
+        }
       } catch (err) {
         handleError(res, err);
       }
@@ -91,74 +111,76 @@ router.get("/whofollows/:currUserId", async (req, res) => {
 router.post("/:currUserId/:targetUserId", async (req, res) => {
     const paramsCheck = checkIdParams(req);
     if (paramsCheck) {
-      handleError(res, `invalid ${paramsCheck} parameter(s)`);
+      handleError(res, `invalid ${paramsCheck} parameter(s)`, 400);
     } else {
       const currUserId = parseInt(req.params.currUserId.trim());
       const targetUserId = parseInt(req.params.targetUserId.trim());
-      const password = req.body.password.trim();
-      let authorized = null;
+      let password, authorized = null;
+      req.body.password
+        ? password = req.body.password.trim()
+        : false;
       try {
-        authorized = await users.authenticateUser(currUserId, password);
+        authorized = await authenticateUser(currUserId, password);
       } catch(err) {
         handleError(res, "error during authentication query");
       }
       if (authorized) {
         try {
-          const response = await createFollow(currUserId, targetUserId);
-          res.json({
-              status: "success",
-              message: "follow created",
-              payload: response
-          });
+          const followExists = await checkFollowExists(currUserId, targetUserId);
+          if (followExists) {
+            handleError(res, 'follow already exists', 403);
+          } else {
+            const response = await createFollow(currUserId, targetUserId);
+            res.json({
+                status: "success",
+                message: "follow created",
+                payload: response
+            });
+          }
         } catch (err) {
           handleError(res, err);
         }
       } else {
-        console.log('Authentication issue')
-        res.status(401);
-        res.json({
-            status: 'fail',
-            message: 'Authentication issue',
-            payload: null,
-        });
+        handleError(res, 'authentication issue', 401);
       }
     }
 });
 
 // deleteFollow: delete follow relationship
-router.patch("/:currUserId/:targetUserId", async (req, res) => {
+router.patch("/delete/:currUserId/:targetUserId", async (req, res) => {
     const paramsCheck = checkIdParams(req);
     if (paramsCheck) {
-      handleError(res, `invalid ${paramsCheck} parameter(s)`);
+      handleError(res, `invalid ${paramsCheck} parameter(s)`, 400);
     } else {
       const currUserId = parseInt(req.params.currUserId.trim());
       const targetUserId = parseInt(req.params.targetUserId.trim());
-      const password = req.body.password.trim();
-      let authorized = null;
+      let password, authorized = null;
+      req.body.password
+        ? password = req.body.password.trim()
+        : false;
       try {
-        authorized = await users.authenticateUser(userId, password);
+        authorized = await authenticateUser(currUserId, password);
       } catch(err) {
         handleError(res, "error during authentication query");
       }
       if (authorized) {
         try {
-          const response = await deleteFollow(currUserId, targetUserId);
-          res.json({
-              status: "success",
-              message: "follow deleted",
-              payload: response
-          });
+          const followExists = await checkFollowExists(currUserId, targetUserId);
+          if (!followExists) {
+            handleError(res, 'follow does not exist', 400);
+          } else {
+            const response = await deleteFollow(currUserId, targetUserId);
+            res.json({
+                status: "success",
+                message: "follow deleted",
+                payload: response
+            });
+          }
         } catch (err) {
           handleError(res, err);
         }
       } else {
-        console.log('Authentication issue')
-        res.status(401);
-        res.json({
-            status: 'fail',
-            message: 'Authentication issue',
-            payload: null,
-        });
+        handleError(res, 'authentication issue', 401);
       }
     }
 });
